fix: Change 게시글, 댓글 RD
완전 삭제 API 수정 - deletedAt 값이 있는 데이터 삭제되도록 조건문 수정
게시글 전체 조회 - 북마크용 전체 조회 API 추가

+++ src/modules/posts/posts.service.js
@@ -69,7 +69,7 @@ const onRequest = async (res, method, pathname, params, key, cb) => {
         pageNum = params.query.pageNum;
       }
 
-      const { postId, annualCategory } = params.query;
+      const { userId, postId, annualCategory } = params.query;
 
       // 게시글 전체 조회
       if (pathname === '/posts') {
@@ -79,6 +79,17 @@ const onRequest = async (res, method, pathname, params, key, cb) => {
             limit: 10,
             offset: (pageNum - 1) * 10,
           });
+          responseData = { result };
+        } catch (err) {
+          responseData = { code: 310 };
+        }
+      }
+
+      // 게시글 전체 조회(북마크 조회용)
+      if (pathname === '/posts' && userId) {
+        try {
+          const { userId } = params.query;
+          const result = await Posts.findAll({ where: { userId } });
           responseData = { result };
         } catch (err) {
           responseData = { code: 310 };
@@ -207,7 +218,7 @@ const onRequest = async (res, method, pathname, params, key, cb) => {
 
     case 'DELETE':
       // 게시글 삭제
-      if (pathname === '/posts') {
+      if (pathname === '/posts' && params.params !== 'admin') {
         try {
           const postId = params.params;
           const findPostData = await Posts.findByPk(postId);
@@ -227,25 +238,20 @@ const onRequest = async (res, method, pathname, params, key, cb) => {
         }
       }
 
-      // 게시글 완전 삭제
-      if (pathname === '/posts/admin') {
+      // 게시글 완전 삭제(게시글이 삭제된지 오래되었을 때 DB 데이터 삭제)
+      if (pathname === '/posts' && params.params === 'admin') {
         try {
           const { contentId } = params.bodies;
-          const findPostData = await Posts.findByPk({ postId: contentId });
 
-          if (!findPostData) {
-            responseData = { code: 372 };
-          } else {
-            await Posts.destroy({ where: { postId: contentId }, force: true });
-            responseData = { code: 371 };
-          }
+          await Posts.destroy({ where: { postId: contentId }, force: true });
+          responseData = { code: 371 };
         } catch (err) {
           responseData = { code: 370 };
         }
       }
 
       // 댓글 삭제
-      if (pathname === '/comments') {
+      if (pathname === '/comments' && params.params !== 'admin') {
         try {
           const { postId } = params.bodies;
           const commentId = params.params;
@@ -264,18 +270,13 @@ const onRequest = async (res, method, pathname, params, key, cb) => {
         }
       }
 
-      // 댓글 완전 삭제
-      if (pathname === '/comments/admin') {
+      // 댓글 완전 삭제(삭제된 댓글이 오래되었을 때 DB 데이터 삭제)
+      if (pathname === '/comments' && params.params === 'admin') {
         try {
           const { contentId } = params.bodies;
-          const findCommentData = await Comments.findByPk({ commentId: contentId });
 
-          if (!findCommentData) {
-            responseData = { code: 452 };
-          } else {
-            await Comments.destroy({ where: { commentId: contentId }, force: true });
-            responseData = { code: 451 };
-          }
+          await Comments.destroy({ where: { commentId: contentId }, force: true });
+          responseData = { code: 451 };
         } catch (err) {
           responseData = { code: 450 };
         }
